Reify the key_span hack as part of the Substitution API.
This... is really nice. The original task was grouping substitutions together so that they can be applied in combination, yielding better changes than we would otherwise have. At the time -- I'm not really sure why offhand, maybe the other API didn't exist yet -- at the time, this was done as a *postprocessing pass* against the stream of Substitution objects from find_iter. This is really unfortunate, since at that point, we've entirely lost AST information. So... we recovered the AST information based on the codepoint spans, and that worked.

But really, any time you lose information and then recover it later is bad, and doing it based off of memorizing the textual spans that represent AST nodes is just the worst.

By putting it *inside* the implementation of find_iter, rather than as a postprocessing pass, we can actually reuse the AST information we discard once we get a Substitution.

(This also has incidental benefits -- the original reason I looked at doing this was actually so that each substitution would know its own key_span for when we have a generic (non-python-specific) CombiningSearcher, which wouldn't know which key_span method to call as it wouldn't know which underlying combined searcher originated the Substitution object.)

PiperOrigin-RevId: 328796281