Add `--version` support.
This change adds `--version` and a version field in the error dump (but doesn't actually uh display that in rxerr_debug yet because I can add it when I need it).

Someone told me that `pkg_resources` is slow, but I couldn't notice any delay, so I did not opt for any shenanigans. (If `pkg_resources` was slow to import or to use, my plan was to, instead of using the `version` action, write a custom action that prints the output of a _callback_ rather than a pre-passed string. We can still do that later if it turns out `pkg_resources` actually is slow in some environments.)

PiperOrigin-RevId: 332886571