Add test for rxerr_debug's argv display.
I actually was doing this so I could use `shlex.join`, but apparently that's 3.8-only, so... that didn't happen.

(Actually, there's no big reason rxerr_debug can't be 3.8-only, other than CI complications -- it isn't user-facing, so it could be written in Common Lisp for all anyone cares. That said, CI complications are a good enough reason to avoid it. What's the point of adding a test if it can't be run in CI?)

PiperOrigin-RevId: 335970363