Fix tests on 3.8-3.9.
The test failure is a testing artifact, and not "real" -- in order to test support for Ellipsis in 3.8+, while retaining Py2-compatibility, I had used `x[...]` instead of just `...`. Unfortunately, the AST for `x[y]` *also* changed in 3.8, and I had neglected to run my test in 3.8!

I don't think refex is going to try to support the AST change for subscript operations, since it doesn't matter nearly as much as constants, so it's fine to just ignore that part and focus on what we really want to support.

The skipIf/2.x support is just legacy boilerplate since this nominally supports 2.7 (but not really, anymore). The 0.2 release of refex will likely remove all of it, because 3.10 is even less backwards compatible with 2.7 than 3.x already was.

PiperOrigin-RevId: 378032095