Speed up performance of `--mode=fix` by 78%.
Refex now only runs matchers if the input has an appropriate type. This is the stupidest possible `--mode=fix` optimization, which is also, coincidentally, part 1 of the coolest possible optimization.

On an ~80KB source file (benchmarking code unfortunately closed source -- it directly tests how refex integrates with Tricorder, because i dummy), before this change, we got the following distribution of runtimes:

```
count    50.000000
mean     16.077780
std       0.052927
min      15.947535
25%      16.044242
50%      16.077291
75%      16.107561
max      16.190649
```

After this change:

```
count    50.000000
mean      3.541441
std       0.045840
min       3.462464
25%       3.499799
50%       3.536258
75%       3.582944
max       3.632229
```

I don't have a great explanation, actually -- really, in the end, the algorithm was `O(n*m*k)`, where `n` is the size of the file, `m` is the number of matchers, and `k` is the depth of the matchers. Here, we've changed it to be, in the best case, `O(n * m)` instead, so I guess that `k` was kind of a big deal.

That, or this has a bug.

---

**Future work:**

The main thing I want to do is actually filter based on type: when we see something of type X, then only even iterate over matchers that can match X. We can see how this works most obviously with, yeah, `AnyOf`: imagine if `AnyOf` actually stored a dict mapping the type to a list of matchers to run when we encounter that type. We could also apply that to fixers for potentially big wins -- most AST nodes would run the matchers for 0 of the fixers, whereas today, every AST node tries a fixer's matcher at least once. This converts the best-case to `O(n)` from `O(n*m)`, and is the change I *actually* wanted to do.

---

Apologies for unrelated formatting changes, I'm testing out a new editor and it's betraying me.

PiperOrigin-RevId: 393974813