Not-yet-ideal example of how to craft the replacement inside of the matcher.
A lot of times, when building matchers, the fact that "what to match" is separated from "what to replace it with" makes it really annoying to craft. A trivial example is anything recursive, really -- rather than building up some absurd combinator thingy, sometimes you just want to walk the AST. Architecturally, Refex actively does not want you to do this. It expects totally separated matchers and replacements. And isn't that... silly?

This change is one step towards that: it really stupidly glues in a place for the matcher to emit its desired replacement.

After this change, it should be pretty straightforward to write matchers by hand which emit their own replacements. Sure, the replacement template and the match template are still separated -- but the replacement template becomes "replace this thing with a variable that it bound to its desired replacement".

---

Future work:

* Make the syntax nicer? There's still a whole lot of gluing together random crap nobody cares about.
* Allow safety checks -- right now, this only works with ShTemplate.
* Add a good error message for what happens when you *don't* use ShTemplate.
* Document this.

---

Far future work:

The fact that "what to match" and "what to replace it with" are specified so far apart has consistently been annoying. I think, going along with this kind of change, we might be able to establish that there is fundamentally no reason they have to be separate: maybe passing `--sub=...` on the command line is actually equivalent to syntax sugar which adds some stuff to the matcher.

So maybe the whole thing should be re-architected so that replacements are *always* specified in the matcher, and there's no second templating layer.

Or maybe not -- like I said, the fact that we can do this kind of change means it really doesn't matter either way, as long as we complete the hookup.

Since I have not been as active as I'd like working on Refex, this remains a bit speculative for now.

PiperOrigin-RevId: 394281721