Validate the context around a `StringMatch` in syntactic templates, rather than giving up.
This doesn't validate that the match itself is "preserved" in structure, but does validate that at least it stands alone within the syntax tree.

A really dumb example of where this sort of falls over is something like `a:b` -- `a:b` would validate, but parse as a different AST, if it were present in a statement context versus a parameter context, but might otherwise structurally not modify the surrounding AST.

I don't know -- it doesn't matter enough to throw up our hands and give up over.

PiperOrigin-RevId: 394289843