Simplify substitution model: `None` --> `{}`.
Before, replacements were either a nonempty dict or `None`. This is, obviously, redundant, so now an empty dict is used instead of `None`.

The end goal of this line of changes is to remove the `Rewriter` class entirely, so that we can more easily refactor where/how the rewrite templates are determined -- in particular, to move the templates into the matcher, for the python matcher search/replaces.

PiperOrigin-RevId: 394565613