Enable pytype for refex.cli.
This involved a silly-looking change to search -- this is described in b/199581484 (bug for PyType requesting this feature). The TL;DR is that PyType doesn't understand `converter`, and the `type` is used for the input parameter. So `attr.ib(type=tuple, converter=tuple)` still won't accept a `list` parameter under pytype, because it is ignoring `converter`. The solution here is to use a common supertype, `Sequence`, which works both as the input and output of the `converter` function and is good enough for our typechecking needs. It works for very few cases, but this is one of them!

PiperOrigin-RevId: 396496033