Make `BasePythonRewritingSearcher.matcher` public.
I tried to do this before because it just seemed like a good idea, but the rewrite to the surrounding module to make things work was painful. However, in this case, I really want it now, so that I can enable pytype on CombiningPythonFixer.

Due to a bug in pytype (Just reported it at b/199587558), an overridden `attr.ib()` is not respected if it's for a protected attribute (`_foo`) and across module boundaries. So, easily fixed by making it public, which was a good idea regardless.

PiperOrigin-RevId: 396496596