RIP in peace: future_string.
This is one of the bits of Python code I'm most proud of in Refex :'(   up there with the syntatic template stuff or the (currently unpublished externally) re_to_re2.

`future_string` was meant to allow straddling Python 2/3 with mixed text/bytes. I forget the details, but fortunately, the details no longer matter: seeing as we're now 100% Python 3, with 100% text, there is no longer any reason for it to exist. And since it can't ever be removed automatically, we must remove it by hand.

This came up a little with the static typing stuff, and at any rate will conflict with the CL chain from static typing stuff, so putting it in the same CL chain.

PiperOrigin-RevId: 396499782