Fix `--recursive` in open source refex release.
Since internally we use a totally different function for getting the file list, it wasn't obvious this didn't work.

Worse, uh, there was not a single unit test anywhere that used `--recursive` or `-R`...

The goal after this is to change the closed-source implementation to fall back to open-source cli.files_from_options() when it would work, which should make closed-source version significantly faster.

PiperOrigin-RevId: 396760097