Suggest `assertEmpty(x)` instead of `assertLen(x, 0)`.
Made slightly complicated by the fact that I have no citation URL.

I saw that there was a lint error for this, and I originally made refex because I literally believe that any lint error that *cna* be automatically fixed *should* be automatically fixed. pylint is evil! So adding it here as well.

PiperOrigin-RevId: 397782396