Fix calls to `assertItemsEqual` -> `assertCountEqual`.
I... didn't really want to do this while Python 2 support existed, because there was no `assertCountEqual` on Python 2, and so this change was not valid. Like many aspects of the Python 2->3 migration, this made it difficult to create safe migration tooling. But now that we're only on Python 3, `assertCountEqual` definitely exists, and where `assertItemsEqual` exists it is a compatibility shim. Hooray!

The Regexp->Regex rewrites were already done in a previous change, so this finishes it off for 2->3 aliases.

PiperOrigin-RevId: 397782683