Simplify statement deletion logic.
Do you know how EXCITED I am for this?

This change was one of the 2.5 motivating use-cases behind changing matchers to return their own replacements. (The other big one was that this allowed for a major optimization.) Since the matchers now return their replacements, rather than it being glued on separately, we can actually modify both the replacement and the thing being replaced side-by-side, either in the searcher (as we do after this CL) or in the matcher itself (TODO in a followup CL.)

Previous to this change, the statement-deletion fix logic was a layering violation: we were trying to use information about the AST after the matcher had already returned the match spans, because we didn't have the replacements available at the same time as the AST. But now, they are both in the same layer, and there's no more awkward code duplication etc.

As an added bonus: not only do we have the AST available for the input, but we *also* have the substitution template available! So we can, for example, do things like check if the replacement was "0 statements", even if it was nonempty (for example, a newline, or a comment). This was not possible (or at least not easy) before. I don't do that in this CL, but it would now be relatively straightforward in a followup.

----

This CL also necessitated making the start/end of token spans optionally exclusive, but that's been a LONG time coming. I thought I had an internal bug for it, but can't find it.

Work to be done in followup CLs:

1) move this at least into the base class python searcher, and hopefully into the matcher itself. The `Once` matcher shows how we can persist state across matcher invocations, exactly like this does by keeping track of deleted nodes -- so there's nothing super impossible here.

2) while we're at it, see how hard it'd be to extend this logic to more general statement rewrites (adding and removing statements, or handling empty replacements that contain a comment, etc.). Who knows. This is very exciting times.

3) Do the same thing for non-statement lists (expr lists, dict-item lists, argument lists, import lists, ...)

---

My excitement is modulated only by the fact that I've got like 3 groups of bugs that have nothing to do with this where users are very upset that refex doesn't quite fit their use case. And instead of fixing real users I'm coming back to year-old refactoring TODOs. I'm sorry.

PiperOrigin-RevId: 397824837