Allow statement deletion from any Python search/replace.
The next step is presumably to get this logic ~somewhere into the matchers and templates themselves, because the searcher should not be a grab bag of every feature! But this is like 99% of the benefit, so woot.

In particular, this allows for statement deletion in fixers now, as the combined fixer inherits from BasePythonRewritingSearcher.

PiperOrigin-RevId: 397841403