Fix `Num()` to only match numbers, not bools, in 3.8+.
Starting in 3.8, the `ast.Num()` AST node went away, but the AST matcher remained. If nothing else, we needed it as a compatibility bridge, so that fixers could work on both 3.7 and 3.8 simultaneously. And so matcher provided the compatibility bridge: on 3.7 and below, it matches `Num` nodes. On 3.8+, it matches `ast.Constant()` nodes which have a number inside them. These occur at the same place in the AST, so it should've been equivalent. However, this implementation had a bug, and incorrectly thought that `ast.Constant(True)` was a number (because `isinstance(True, int)`).

By fixing the `Num` matcher, we also fix the `x is 42` fixer, which was incorrectly flagging `x is True` checks. Unlike `x is 42` checks, `is True` checks are not buggy at all!

PiperOrigin-RevId: 450587598