31 is number 2: add Python 3.12 support to Refex.
The way in which expansion patterns are parsed was changed in 3.12 to lean even more on the variable names being stored in the parsed regular expression. So we had to change the fake parsed regex to write down group names.

This is piling hacks on top of hacks, likely support for this with non-regexps should be removed entirely, so that we always can have a real pattern available.

PiperOrigin-RevId: 712583753